requirements.txt
streamlit
pandas
numpy
seaborn
matplotlib
wordcloud
textblob
scikit-learn
openai
python-dotenv
sqlalchemy
langchain
langchain-openai
langchain-experimental
langchain-community
tiktoken


 
